Split search tests
